requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
python-dateutil==2.9.0
jinja2==3.1.4
